backend/services/satellite_service.py
"""
Satellite Data Service - SentinelHub API Integration
Handles satellite imagery retrieval and band data processing
"""

import os
import requests
from datetime import datetime, timedelta
from typing import Dict, List, Optional, Tuple
import numpy as np
from dotenv import load_dotenv
import base64
import io
from PIL import Image

load_dotenv()

class SatelliteService:
    """Service for interacting with SentinelHub API"""
    
    def __init__(self):
        self.client_id = os.getenv('SENTINELHUB_CLIENT_ID')
        self.client_secret = os.getenv('SENTINELHUB_CLIENT_SECRET')
        self.instance_id = os.getenv('SENTINELHUB_INSTANCE_ID')
        self.base_url = "https://services.sentinel-hub.com"
        self.token = None
        self.token_expiry = None
        
    def _get_access_token(self) -> str:
        """Get OAuth2 access token from SentinelHub"""
        if self.token and self.token_expiry and datetime.now() < self.token_expiry:
            return self.token
            
        url = f"{self.base_url}/oauth/token"
        data = {
            'grant_type': 'client_credentials',
            'client_id': self.client_id,
            'client_secret': self.client_secret
        }
        
        try:
            response = requests.post(url, data=data)
            response.raise_for_status()
            token_data = response.json()
            
            self.token = token_data['access_token']
            # Set expiry 5 minutes before actual expiry for safety
            self.token_expiry = datetime.now() + timedelta(seconds=token_data['expires_in'] - 300)
            
            return self.token
        except Exception as e:
            raise Exception(f"Failed to get SentinelHub access token: {str(e)}")
    
    def get_satellite_data(
        self,
        bbox: List[float],
        start_date: str,
        end_date: str,
        max_cloud_coverage: float = 20.0
    ) -> Dict:
        """
        Retrieve satellite data for a bounding box
        
        Args:
            bbox: [min_lon, min_lat, max_lon, max_lat]
            start_date: ISO format date string (YYYY-MM-DD)
            end_date: ISO format date string (YYYY-MM-DD)
            max_cloud_coverage: Maximum acceptable cloud coverage percentage
            
        Returns:
            Dictionary containing NDVI, NDMI, and metadata
        """
        token = self._get_access_token()
        
        # Evalscript for calculating NDVI and NDMI
        evalscript = """
        //VERSION=3
        function setup() {
          return {
            input: [{
              bands: ["B04", "B08", "B11", "SCL"],
              units: "DN"
            }],
            output: [
              {
                id: "ndvi",
                bands: 1,
                sampleType: "FLOAT32"
              },
              {
                id: "ndmi",
                bands: 1,
                sampleType: "FLOAT32"
              },
              {
                id: "rgb",
                bands: 3,
                sampleType: "AUTO"
              }
            ]
          };
        }

        function evaluatePixel(sample) {
          // Calculate NDVI: (NIR - RED) / (NIR + RED)
          let ndvi = (sample.B08 - sample.B04) / (sample.B08 + sample.B04);
          
          // Calculate NDMI: (NIR - SWIR) / (NIR + SWIR)
          let ndmi = (sample.B08 - sample.B11) / (sample.B08 + sample.B11);
          
          // Mask clouds (SCL classification)
          if (sample.SCL === 3 || sample.SCL === 8 || sample.SCL === 9) {
            ndvi = -999;
            ndmi = -999;
          }
          
          return {
            ndvi: [ndvi],
            ndmi: [ndmi],
            rgb: [sample.B04 / 3000, sample.B08 / 3000, sample.B04 / 3000]
          };
        }
        """
        
        # Request payload
        payload = {
            "input": {
                "bounds": {
                    "bbox": bbox,
                    "properties": {
                        "crs": "http://www.opengis.net/def/crs/EPSG/0/4326"
                    }
                },
                "data": [{
                    "type": "sentinel-2-l2a",
                    "dataFilter": {
                        "timeRange": {
                            "from": f"{start_date}T00:00:00Z",
                            "to": f"{end_date}T23:59:59Z"
                        },
                        "maxCloudCoverage": max_cloud_coverage
                    }
                }]
            },
            "output": {
                "width": 512,
                "height": 512,
                "responses": [
                    {
                        "identifier": "ndvi",
                        "format": {"type": "image/tiff"}
                    },
                    {
                        "identifier": "ndmi",
                        "format": {"type": "image/tiff"}
                    }
                ]
            },
            "evalscript": evalscript
        }
        
        headers = {
            'Authorization': f'Bearer {token}',
            'Content-Type': 'application/json'
        }
        
        url = f"{self.base_url}/api/v1/process"
        
        try:
            response = requests.post(url, json=payload, headers=headers)
            response.raise_for_status()
            
            # Process the response
            return self._process_satellite_response(response.content, bbox)
            
        except requests.exceptions.HTTPError as e:
            if e.response.status_code == 400:
                raise Exception(f"No satellite data available for the specified date range and cloud coverage")
            raise Exception(f"SentinelHub API error: {str(e)}")
        except Exception as e:
            raise Exception(f"Failed to retrieve satellite data: {str(e)}")
    
    def _process_satellite_response(self, response_content: bytes, bbox: List[float]) -> Dict:
        """Process the satellite API response and extract statistics"""
        # In a real implementation, you would parse the TIFF data
        # For now, we'll return a structured response
        
        # This is a simplified version - in production, use rasterio to read TIFF
        return {
            "status": "success",
            "bbox": bbox,
            "acquisition_date": datetime.now().isoformat(),
            "cloud_coverage": 5.2,
            "data_available": True
        }
    
    def get_ndvi_timeseries(
        self,
        bbox: List[float],
        start_date: str,
        end_date: str,
        interval_days: int = 5
    ) -> List[Dict]:
        """
        Get NDVI time series data for a farm boundary
        
        Args:
            bbox: Bounding box coordinates
            start_date: Start date (YYYY-MM-DD)
            end_date: End date (YYYY-MM-DD)
            interval_days: Interval between data points
            
        Returns:
            List of NDVI measurements with timestamps
        """
        token = self._get_access_token()
        
        # Statistical API for time series
        evalscript = """
        //VERSION=3
        function setup() {
          return {
            input: [{
              bands: ["B04", "B08", "SCL"],
              units: "DN"
            }],
            output: [
              {
                id: "ndvi_stats",
                bands: 1,
                sampleType: "FLOAT32"
              },
              {
                id: "dataMask",
                bands: 1
              }
            ]
          };
        }

        function evaluatePixel(sample) {
          let ndvi = (sample.B08 - sample.B04) / (sample.B08 + sample.B04);
          
          // Mask clouds
          let mask = 1;
          if (sample.SCL === 3 || sample.SCL === 8 || sample.SCL === 9) {
            mask = 0;
          }
          
          return {
            ndvi_stats: [ndvi],
            dataMask: [mask]
          };
        }
        """
        
        payload = {
            "input": {
                "bounds": {
                    "bbox": bbox,
                    "properties": {
                        "crs": "http://www.opengis.net/def/crs/EPSG/0/4326"
                    }
                },
                "data": [{
                    "type": "sentinel-2-l2a",
                    "dataFilter": {
                        "timeRange": {
                            "from": f"{start_date}T00:00:00Z",
                            "to": f"{end_date}T23:59:59Z"
                        }
                    }
                }]
            },
            "aggregation": {
                "timeRange": {
                    "from": f"{start_date}T00:00:00Z",
                    "to": f"{end_date}T23:59:59Z"
                },
                "aggregationInterval": {
                    "of": f"P{interval_days}D"
                },
                "evalscript": evalscript,
                "resx": 10,
                "resy": 10
            },
            "calculations": {
                "ndvi_stats": {
                    "statistics": {
                        "default": {
                            "percentiles": {
                                "k": [25, 50, 75]
                            }
                        }
                    }
                }
            }
        }
        
        headers = {
            'Authorization': f'Bearer {token}',
            'Content-Type': 'application/json'
        }
        
        url = f"{self.base_url}/api/v1/statistics"
        
        try:
            response = requests.post(url, json=payload, headers=headers)
            response.raise_for_status()
            
            data = response.json()
            return self._parse_timeseries_response(data)
            
        except Exception as e:
            raise Exception(f"Failed to retrieve NDVI time series: {str(e)}")
    
    def _parse_timeseries_response(self, response_data: Dict) -> List[Dict]:
        """Parse the statistical API response into a clean time series"""
        timeseries = []
        
        for interval in response_data.get('data', []):
            stats = interval.get('outputs', {}).get('ndvi_stats', {}).get('bands', {}).get('B0', {}).get('stats', {})
            
            timeseries.append({
                'date': interval.get('interval', {}).get('from', '').split('T')[0],
                'ndvi_mean': stats.get('mean'),
                'ndvi_min': stats.get('min'),
                'ndvi_max': stats.get('max'),
                'ndvi_std': stats.get('stDev'),
                'percentile_25': stats.get('percentiles', {}).get('25.0'),
                'percentile_50': stats.get('percentiles', {}).get('50.0'),
                'percentile_75': stats.get('percentiles', {}).get('75.0')
            })
        
        return timeseries
    
    def calculate_bbox_from_polygon(self, coordinates: List[List[float]]) -> List[float]:
        """
        Calculate bounding box from polygon coordinates
        
        Args:
            coordinates: List of [lon, lat] pairs
            
        Returns:
            [min_lon, min_lat, max_lon, max_lat]
        """
        lons = [coord[0] for coord in coordinates]
        lats = [coord[1] for coord in coordinates]
        
        return [min(lons), min(lats), max(lons), max(lats)]
