backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
scikit-learn==1.5.2
pandas==2.2.3
numpy<2.0.0
python-multipart==0.0.20
joblib==1.4.2
google-generativeai
python-dotenv
twilio
torch
torchvision
Pillow>=10.0.0

# Satellite Analytics Dependencies
sentinelhub>=3.10.0
rasterio>=1.3.0
shapely>=2.0.0
firebase-admin>=6.0.0
APScheduler>=3.10.0
